Clean up and Fix
An issue arose during testing where if the user did not have any pinned projects their "pinned" projects would be "popular" projects which have a different html node structure to them. I then changed it to allow the popular projects to be passed through instead of the pinned projects but added a type to GitHubInfo object so people can see what is being passed through.

I also cleaned up some code that was not needed and removed the template files (for weather api).

TO DO:
- Host on azure
- Get code review